Refactor Contact Us modal and related components
- Split the Contact Us functionality into multiple components: `ContactUsHeader`, `ContactUsFields`, `ContactUsFooter`, `ContactUsForm`, and `ContactUsModal` for improved organization and maintainability.
- Updated the `ContactUsModal` to utilize the new components, enhancing readability and separation of concerns.
- Introduced a new `ContactUsModalRouter` for managing modal state and rendering.
- Improved form handling and error management within the new structure.
