+++ apps/next/components/website/contact-us/ContactUsHeader.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+export function ContactUsHeader() {
+  return (
+    <div className="flex flex-col">
+      <h3 className="text-lg font-medium">Contact Us</h3>
+      <p className="text-sm text-default-500">
+        Fill out the form below, and we&apos;ll get back to you soon.
+      </p>
+    </div>
+  );
+}
